Improve text color, light/dark mode handling
The previous mix-blend-mode approach was unstable, and led to some ugly
results with text over images when they weren't an even gray. It looks a
lot better if we use rgba values.

This change introduces two sets of colors (dark/light text): setting
them by default in light and dark mode, and making them configurable on
a per-image basis in the gallery (for full images).